relocating webpack files
