+++ package.json
@@ -25,13 +25,13 @@
   "scripts": {
     "clean": "rimraf dist build",
     "prebuild": "rimraf dist",
-    "build": "webpack --mode production --config webpack.library.cfg.js --display-modules",
+    "build": "webpack --mode production --config webpack/library.cfg.js --display-modules",
     "build:light": "babel ./src --out-dir ./dist",
-    "build:examples": "webpack --mode production --config webpack.examples.cfg.js",
+    "build:examples": "webpack --mode production --config webpack/examples.cfg.js",
     "prepublish": "npm run test && npm run build",
     "lint": "eslint ./src",
     "lint:fix": "eslint --fix ./src",
-    "start": "node webpack.server.js",
+    "start": "node webpack/server.js",
     "start:debug": "weinre --boundHost `my-local-ip` ",
     "test": "jest --coverage",
     "update": "ncu -a --packageFile package.json",
